+++ NEWS.md
@@ -52,7 +52,22 @@ This affects `derive_var_dthcaus()`. (#1727)
 - The `date` field of `event_source()` and `censor_source()` accepts expressions
 now. This affects `derive_param_tte()`. (#1727)
 
+
 ## Breaking Changes
+
+- `create_query_data()` and `derive_vars_query()` updated to rename variables in 
+    query data set as follows:
+
+    - `VAR_PREFIX` to `PREFIX`
+    - `QUERY_NAME` to `GRPNAME`
+    - `QUERY_ID` to `GRPID`
+    - `QUERY_SCOPE` to `SCOPE`
+    - `QUERY_SCOPE_NUM` to `SCOPEN`
+    - `TERM_LEVEL` to `SRCVAR`
+    - `TERM_NAME` to `TERMNAME`
+    - `TERM_ID` to `TERMID`
+  
+    Users need to adjust their `get_terms()` function accordingly.
 
 - The `aval_fun` argument of `derive_param_exist_flag()` was deprecated in favor
 of the `set_values_to` argument. (#1727)
@@ -107,6 +122,8 @@ as well. (#1694)
 - Examples for `derive_var_extreme_flag()` were reworked to reduce runtime that occasionally led to failing CI check. (#1780)
 
 - `create_period_dataset()` had a bug that led to an error when both DT and DTM columns existed. (#1845)
+
+- External functions are now consistently imported via namespace. `package::function()` calls have been removed from `admiral` functions. (#1842)
 
 # admiral 0.10.2